Updated GrailsProfilePublishGradlePlugin and GrailsCentralPublishGradlePlugin
