+++ build.gradle
@@ -81,6 +81,8 @@ dependencies {
     documentation "org.codehaus.groovy:groovy-cli-picocli:$groovyVersion"
 
     compile gradleApi()
+    compile "io.github.gradle-nexus:publish-plugin:1.0.0"
+    compile "com.bmuschko:gradle-nexus-plugin:2.3.1"
     compile "org.grails:grails-gradle-model:$grailsVersion"
     compile "org.grails:grails-shell:$grailsVersion", {
         exclude group: "org.slf4j", module: "slf4j-simple"
