--- src/main/resources/META-INF/gradle-plugins/org.grails.grails-profile-publish.properties
@@ -1,1 +1,0 @@
-implementation-class=org.grails.gradle.plugin.profiles.GrailsProfilePublishGradlePlugin
