+++ src/main/resources/META-INF/gradle-plugins/org.grails.internal.grails-plugin-publish.properties
@@ -1,0 +1,1 @@
+implementation-class=org.grails.gradle.plugin.publishing.internal.GrailsCentralPublishGradlePlugin
